+++ tests/LoanApplicationConsoleTests/LoanApplicationRiskDeciderTests.cs
@@ -27,7 +27,7 @@ public class LoanApplicationRiskDeciderTests
     [InlineData(1_000_000, 60, 950)]
     [InlineData(1_000_000, 60, 951)]
     [InlineData(1_000_000, 59, 950)]
-    public void LoanAmount_OneMillionOrGreater_Approved(int loanAmount, short loanToValueRate, short creditScore)
+    public void LoanAmount_OneMillionOrGreater_Approved(int loanAmount, byte loanToValueRate, short creditScore)
     {
         var applicationStatus = RunRiskDecider(loanAmount, loanToValueRate, creditScore);
         
@@ -41,7 +41,7 @@ public class LoanApplicationRiskDeciderTests
     [InlineData(1_000_000, 61, 950)]
     [InlineData(1_000_000, 60, 949)]
     [InlineData(1_000_000, 61, 949)]
-    public void LoanAmount_OneMillionOrGreater_Declined(int loanAmount, short loanToValueRate, short creditScore)
+    public void LoanAmount_OneMillionOrGreater_Declined(int loanAmount, byte loanToValueRate, short creditScore)
     {
         var status = RunRiskDecider(loanAmount, loanToValueRate, creditScore);
         
@@ -54,7 +54,7 @@ public class LoanApplicationRiskDeciderTests
     [InlineData(999_999, 79, 800)]
     [InlineData(999_999, 80, 900)]
     [InlineData(999_999, 89, 900)]
-    public void LoanAmount_LessThanOneMillion_Approved(int loanAmount, short loanToValueRate, short creditScore)
+    public void LoanAmount_LessThanOneMillion_Approved(int loanAmount, byte loanToValueRate, short creditScore)
     {
         var status = RunRiskDecider(loanAmount, loanToValueRate, creditScore);
         
@@ -66,14 +66,14 @@ public class LoanApplicationRiskDeciderTests
     [InlineData(999_999, 79, 799)]
     [InlineData(999_999, 89, 899)]
     [InlineData(999_999, 90, 999)]
-    public void LoanAmount_LessThanOneMillion_Declined(int loanAmount, short loanToValueRate, short creditScore)
+    public void LoanAmount_LessThanOneMillion_Declined(int loanAmount, byte loanToValueRate, short creditScore)
     {
         var status = RunRiskDecider(loanAmount, loanToValueRate, creditScore);
         
         Assert.True(status is LoanApplicationStatus.Declined);
     }
 
-    private static LoanApplicationStatus RunRiskDecider(int loanAmount, short loanToValueRate, short creditScore)
+    private static LoanApplicationStatus RunRiskDecider(int loanAmount, byte loanToValueRate, short creditScore)
     {
         LoanApplicationRiskDecider riskDecider = new();
